# docs/swagger.yaml
basePath: /
definitions:
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  service.CreateSubscriptionDTO:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing subscriptions
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: List subscriptions filtered by user_id and optionally service_name
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription DTO
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/service.CreateSubscriptionDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get a subscription by its ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription DTO
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/service.CreateSubscriptionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      consumes:
      - application/json
      description: Calculate total subscription price for a user in a period, optionally
        filtered by service_name
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      - description: From month-year MM-YYYY
        in: query
        name: from
        required: true
        type: string
      - description: To month-year MM-YYYY
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sum subscriptions
      tags:
      - subscriptions
swagger: "2.0"